feat: use ooni/probe-cli 3.13.0 and expose dnscheck
This diff uses ooni/probe-cli 3.13.0. I tested that this
specific version was working in linux/amd64.

While there, I noticed that dnscheck was not exposed,
for understandable reasons (it has been broken until the
release of 3.13.0) and exposed it. (It has been quite a
pleasing experience to be able to expose it by just
adding a single line of code for declaring it, well done!)

Reference issue https://github.com/ooni/probe/issues/1843