chore: use ooni/probe-cli@v3.19.0-alpha.4
This is the latest alpha release as of this writing and contains
the necessary code to run riseupvpn as a default-disable experimental
nettest. As documented by https://github.com/ooni/probe/issues/2576,
this release also has an issue when running that we'll fix on the
probe-cli side as soon as possible.